Refactors __main__ scripts | moves most code (#25)
- force pushed missing v1 in fastasi imports following renaming
  from fast_ai_1 to fast_ai_v1

* Refactors package __main__ | code in new scripts

- creates two new scripts: plots & extractions
- distributes procedural and argparse code from src.__main__ between
  the two new scripts and utilities

* Refactors fai __main__ | moves code to fai utilities

- moves procedural and argparse code from src.pipelines.fast_ai_1.__main__ to fai utilities
  script

* Reformats code

* Adds DataBunch comment header

* Renames fast_ai_1 to fast_ai_v1

- the new name mores explicit the meaning of the "1" in the directory name
